+++ packages/types/src/codec/Compact.ts
@@ -2,15 +2,17 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { AnyNumber, Codec, Constructor, ICompact, InterfaceTypes, Registry } from '../types';
+import { H256 } from '@polkadot/types/interfaces';
+import { AnyJson, AnyNumber, Codec, Constructor, ICompact, InterfaceTypes, Registry } from '../types';
 
 import BN from 'bn.js';
 import { compactAddLength, compactFromU8a, compactStripLength, compactToU8a, isBigInt, isBn, isNumber, isString } from '@polkadot/util';
 import { DEFAULT_BITLENGTH } from '@polkadot/util/compact/defaults';
+import { blake2AsU8a } from '@polkadot/util-crypto';
 
 import typeToConstructor from './utils/typeToConstructor';
 import { UIntBitLength } from './AbstractInt';
-import Base from './Base';
+import Raw from './Raw';
 
 export interface CompactEncodable extends Codec {
   bitLength (): number;
@@ -26,17 +28,17 @@ export interface CompactEncodable extends Codec {
  * used by other types to add length-prefixed encoding, or in the case of wrapped types, taking
  * a number and making the compact representation thereof
  */
-export default class Compact<T extends CompactEncodable> extends Base<T> implements ICompact<T> {
-  private readonly _Type: Constructor<T>;
+export default class Compact<T extends CompactEncodable> implements ICompact<T> {
+  public readonly registry: Registry;
+
+  readonly #Type: Constructor<T>;
+
+  readonly #raw: T;
 
   constructor (registry: Registry, Type: Constructor<T> | keyof InterfaceTypes, value: Compact<T> | AnyNumber = 0) {
-    const _Type = typeToConstructor(registry, Type);
-
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    super(registry, Compact.decodeCompact<T>(registry, _Type, value));
-
-    this._Type = _Type;
+    this.registry = registry;
+    this.#Type = typeToConstructor(registry, Type);
+    this.#raw = Compact.decodeCompact<T>(registry, this.#Type, value) as T;
   }
 
   public static with<T extends CompactEncodable> (Type: Constructor<T> | keyof InterfaceTypes): Constructor<Compact<T>> {
@@ -67,7 +69,7 @@ export default class Compact<T extends CompactEncodable> extends Base<T> impleme
   /** @internal */
   public static decodeCompact<T extends CompactEncodable> (registry: Registry, Type: Constructor<T>, value: Compact<T> | AnyNumber): CompactEncodable {
     if (value instanceof Compact) {
-      return new Type(registry, value._raw);
+      return new Type(registry, value.#raw);
     } else if (isString(value) || isNumber(value) || isBn(value) || isBigInt(value)) {
       return new Type(registry, value);
     }
@@ -78,19 +80,40 @@ export default class Compact<T extends CompactEncodable> extends Base<T> impleme
   }
 
   /**
+   * @description The length of the value when encoded as a Uint8Array
+   */
+  public get encodedLength (): number {
+    return this.toU8a().length;
+  }
+
+  /**
+   * @description returns a hash of the contents
+   */
+  public get hash (): H256 {
+    return new Raw(this.registry, blake2AsU8a(this.toU8a(), 256));
+  }
+
+  /**
+   * @description Checks if the value is an empty value
+   */
+  public get isEmpty (): boolean {
+    return this.#raw.isEmpty;
+  }
+
+  /**
    * @description Returns the number of bits in the value
    */
   public bitLength (): number {
-    return this._raw.bitLength();
+    return this.#raw.bitLength();
   }
 
   /**
    * @description Compares the value of the input to see if there is a match
    */
   public eq (other?: unknown): boolean {
-    return this._raw.eq(
+    return this.#raw.eq(
       other instanceof Compact
-        ? other._raw
+        ? other.#raw
         : other
     );
   }
@@ -99,21 +122,49 @@ export default class Compact<T extends CompactEncodable> extends Base<T> impleme
    * @description Returns the BN representation of the number
    */
   public toBn (): BN {
-    return this._raw.toBn();
+    return this.#raw.toBn();
+  }
+
+  /**
+   * @description Returns a hex string representation of the value. isLe returns a LE (number-only) representation
+   */
+  public toHex (isLe?: boolean): string {
+    return this.#raw.toHex(isLe);
+  }
+
+  /**
+   * @description Converts the Object to to a human-friendly JSON, with additional fields, expansion and formatting of information
+   */
+  public toHuman (isExtended?: boolean): AnyJson {
+    return this.#raw.toHuman(isExtended);
+  }
+
+  /**
+   * @description Converts the Object to JSON, typically used for RPC transfers
+   */
+  public toJSON (): AnyJson {
+    return this.#raw.toJSON();
   }
 
   /**
    * @description Returns the number representation for the value
    */
   public toNumber (): number {
-    return this._raw.toNumber();
+    return this.#raw.toNumber();
   }
 
   /**
    * @description Returns the base runtime type name for this instance
    */
   public toRawType (): string {
-    return `Compact<${this.registry.getClassName(this._Type) || this._raw.toRawType()}>`;
+    return `Compact<${this.registry.getClassName(this.#Type) || this.#raw.toRawType()}>`;
+  }
+
+  /**
+   * @description Returns the string representation of the value
+   */
+  public toString (): string {
+    return this.#raw.toString();
   }
 
   /**
@@ -122,13 +173,13 @@ export default class Compact<T extends CompactEncodable> extends Base<T> impleme
    */
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   public toU8a (isBare?: boolean): Uint8Array {
-    return Compact.encodeU8a(this._raw.toBn());
+    return Compact.encodeU8a(this.#raw.toBn());
   }
 
   /**
    * @description Returns the embedded [[UInt]] or [[Moment]] value
    */
   public unwrap (): T {
-    return this._raw;
+    return this.#raw;
   }
 }